+++ combinators/iteration/many.ts
@@ -1,5 +1,5 @@
 import { alt } from "../alternation/alt.ts";
-import type {  Parser } from "../../index.ts";
+import type { Parser } from "../../index.ts";
 import { result } from "../../primitives/result.ts";
 
 /**
