+++ combinators/iteration/many1.ts
@@ -1,5 +1,5 @@
 import { many } from "./many.ts";
-import type {  Parser } from "../../index.ts";
+import type { Parser } from "../../index.ts";
 import { result } from "../../primitives/result.ts";
 
 /**
