+++ primitives/mod.ts
@@ -5,3 +5,4 @@
  */
 
 export { result } from "./result.ts";
+export { zero } from "./zero.ts";
